1) cambio en la base de datos, se agrega el campo "finalizada" para detectar si la oferta esta finalizada o no 2) Agrupamiento de informacion y opciones_basicas de cada oferta/carrera/evento en el boton "Más" 3) Filtro por año y nombre de oferta, y Odenamiento por los mismos campos en la vista de Ofertas/Carreras/eventos. 4) En Ofertas/index carga la primera vez en el tab Carreras, y luego se carga el tab según la accion donde estaba parado, o sea, vuelve al tab correspondiente (Of/Carr/Ev) 5) Opciones de Finalizar Oferta: no se puede borrar ni editar, y no se puede realizar ningun cambio en los inscriptos de la Oferta finalizada, solo ver y descargar los datos en excel/pdf. 6) Se cambio el tipo de letra para los certificados, ahora esta utilizando Segoe UI (falta arreglar los mismos segun este ultimo cambio - tamaño, bold, interlineado, etc).

+++ app/views/inscripciones/carreras/index.blade.php
@@ -35,9 +35,10 @@
     @endif -->
     <div>
         <a class='btn btn-primary' href="{{ URL::route('ofertas.index') }}" title="Volver al listado de Ofertas" >Volver</a>
-        @if(sizeof($preinscripciones))
+        @if((sizeof($preinscripciones))&&(!$oferta->estaFinalizada()))
             {{ Form::open(array('class' => 'confirm-delete', 'style' => 'display: inline-block;', 'method' => 'DELETE', 'route' => array('ofertas.inscripciones.limpiar', $oferta->id))) }}
-                {{ Form::submit('Limpiar Carrera', array('class' => 'btn btn-danger','title'=>'Eliminar todos los preinscriptos de la Carrera')) }}
+                <input id='mjeBorrar' value="¿Está seguro que desea borrar todos los preinscriptos a esta Oferta?" type="hidden" />
+                {{ Form::submit('Borrar inscriptos de Carrera', array('class' => 'btn btn-danger','title'=>'Eliminar todos los preinscriptos de la Carrera')) }}
             {{ Form::close() }}
         @endif
     </div>
